first version is the hardest
The initial structure and format of the website is
the most critical aspect, so lets take time to get
it just right. The idea is not to overdo but come up
with a simple writing structure.

This version has the template and an initial story
about not going uninvited, while using the ever famous
lord Shiva, Sati and Daksh Prajapati discourse.