change z-index to auto for order of image
The hero image was coming on top of the header
text, so changing the z-level to auto fixed it.
Read https://developer.mozilla.org/en-US/docs/Web/CSS/z-index
for more details, but it appears that the order of the
content added in css determines the order in which
the content is placed in a specific place.